+++ src/base/pages/Landing.vue
@@ -83,7 +83,6 @@
 
       <GroupGalleryCards
         :groups="groupsToShow"
-        :is-logged-in="false"
         @preview="preview(arguments[0])"
       />
 
@@ -110,9 +109,8 @@
       >
         <a
           slot="code"
-          v-t="'ABOUT_KARROT.LINKS.CODE'"
           href="https://github.com/yunity/karrot-frontend"
-        />
+        >GitHub</a>
       </i18n>
     </section>
 
@@ -133,9 +131,8 @@
             >
               <a
                 slot="code"
-                v-t="'ABOUT_KARROT.LINKS.CODE'"
                 href="https://github.com/yunity/karrot-frontend"
-              />
+              >GitHub</a>
               <a
                 slot="forum"
                 v-t="'ABOUT_KARROT.LINKS.FORUM'"
@@ -148,9 +145,8 @@
               />
               <a
                 slot="translations"
-                v-t="'ABOUT_KARROT.LINKS.TRANSLATIONS'"
                 href="https://www.transifex.com/yunity-1/karrot/frontend/"
-              />
+              >Transifex</a>
             </i18n>
           </p>
         </li>