+++ src/groupInfo/components/GroupGalleryCard.story.js
@@ -1,18 +1,10 @@
-import { createDatastore, storybookDefaults as defaults } from '>/helpers'
+import { storybookDefaults as defaults } from '>/helpers'
 import { storiesOf } from '@storybook/vue'
 import { action } from '@storybook/addon-actions'
 
 import GroupGalleryCard from './GroupGalleryCard'
 
 import { groupsMock } from '>/mockdata'
-
-const datastore = createDatastore({
-  applications: {
-    getters: {
-      getMineInGroup: () => () => null,
-    },
-  },
-})
 
 const methods = {
   preview: action('view group preview'),
@@ -30,7 +22,6 @@ storiesOf('GroupGalleryCard', module)
       },
       on: methods,
     }),
-    store: datastore,
   }))
   .add('isMember = false', () => defaults({
     render: h => h(GroupGalleryCard, {
@@ -42,7 +33,18 @@ storiesOf('GroupGalleryCard', module)
       },
       on: methods,
     }),
-    store: datastore,
+  }))
+  .add('isMember = false, application pending', () => defaults({
+    render: h => h(GroupGalleryCard, {
+      props: {
+        group: {
+          ...groupsMock[0],
+          isMember: false,
+          myApplicationPending: true,
+        },
+      },
+      on: methods,
+    }),
   }))
   .add('without public description', () => defaults({
     render: h => h(GroupGalleryCard, {
@@ -55,5 +57,4 @@ storiesOf('GroupGalleryCard', module)
       },
       on: methods,
     }),
-    store: datastore,
   }))