+++ CHANGELOG.md
@@ -21,13 +21,30 @@ Please document your changes in this format:
 ```
 
 ## [Unreleased]
+
+## [8.5.1] - 2020-05-24
+### Changed
+- Show ongoing pickups to all group members, not just those who signed up [#1789] @tiltec
+- Updated translations
+
+## [8.5.0] - 2020-05-18
+### Added
+- Notification when a new Karrot version is available @tiltec
+
 ### Changed
 - Subscribe everybody to conflict resolution emails by default @nicksellen
 - Speed up page load (hopefully) @tiltec @nicksellen
-- Hide buttons in group sidenav, change notification icons, added settings to latest messages menu [ #1987 @brnsolikyl @nicksellen
+- Hide buttons in group sidenav, change notification icons, added settings to latest messages menu [#1987] @brnsolikyl @nicksellen
+- Show instruction on group preview page how to revive groups without members @tiltec
+- Switch to group when clicking on user-related notificiations @tiltec
+- Remove "public description" toggle from group description page, as it is available from the sidenav already @tiltec
+- In group gallery, move groups with your application pending to the top @tiltec
 
 ### Fixed
 - Unsubscribe link in pickup notification email did not work @nicksellen
+- Names in application chat were normally not visible for the applicant @tiltec
+- "Forgot your password?" page didn't show validation errors @tiltec
+- Selecting to join playground group after signup didn't actually join the group @tiltec
 
 ## [8.4.0] - 2020-04-25
 ### Added
@@ -529,10 +546,14 @@ After months of planning and programming, the first version of ~the foodsaving t
 [#1509]: https://github.com/yunity/karrot-frontend/issues/1509
 [#1690]: https://github.com/yunity/karrot-frontend/issues/1690
 [#1691]: https://github.com/yunity/karrot-frontend/issues/1691
+[#1789]: https://github.com/yunity/karrot-frontend/issues/1789
 [#1799]: https://github.com/yunity/karrot-frontend/issues/1799
 [#1982]: https://github.com/yunity/karrot-frontend/issues/1982
+[#1987]: https://github.com/yunity/karrot-frontend/issues/1987
 
-[Unreleased]: https://github.com/yunity/karrot-frontend/compare/v8.4.0...HEAD
+[Unreleased]: https://github.com/yunity/karrot-frontend/compare/v8.5.1...HEAD
+[8.5.1]: https://github.com/yunity/karrot-frontend/compare/v8.5.0...v8.5.1
+[8.5.0]: https://github.com/yunity/karrot-frontend/compare/v8.4.0...v8.5.0
 [8.4.0]: https://github.com/yunity/karrot-frontend/compare/v8.3.0...v8.4.0
 [8.3.0]: https://github.com/yunity/karrot-frontend/compare/v8.2.0...v8.3.0
 [8.2.0]: https://github.com/yunity/karrot-frontend/compare/v8.1.2...v8.2.0